+++ src/game/client/sdk/hud/sdk_hud_armor.cpp
@@ -1,0 +1,128 @@
+//========= Copyright © 1996-2008, Valve Corporation, All rights reserved. ============//
+//
+// Purpose: 
+//
+// $NoKeywords: $
+//
+//=====================================================================================//
+
+#include "cbase.h"
+#include "hud.h"
+#include "hudelement.h"
+#include "hud_macros.h"
+#include "hud_numericdisplay.h"
+#include "iclientmode.h"
+
+#include "vgui_controls/AnimationController.h"
+#include "vgui/ILocalize.h"
+
+#include "c_sdk_player.h"
+// memdbgon must be the last include file in a .cpp file!!!
+#include "tier0/memdbgon.h"
+
+#define INIT_ARMOR	-1
+
+//-----------------------------------------------------------------------------
+// Purpose: Displays suit power (armor) on hud
+//-----------------------------------------------------------------------------
+class CSDKHudArmor : public CHudElement, public CHudNumericDisplay
+{
+	DECLARE_CLASS_SIMPLE( CSDKHudArmor, CHudNumericDisplay );
+
+public:
+	CSDKHudArmor( const char *pElementName );
+	void Init( void );
+	void Reset( void );
+	void VidInit( void );
+	void OnThink( void );
+	
+private:
+	int		m_iArmor;	
+};
+
+DECLARE_HUDELEMENT( CSDKHudArmor );
+
+//-----------------------------------------------------------------------------
+// Purpose: Constructor
+//-----------------------------------------------------------------------------
+CSDKHudArmor::CSDKHudArmor( const char *pElementName ) : CHudElement( pElementName ), CHudNumericDisplay(NULL, "HudSuit")
+{
+	SetHiddenBits( HIDEHUD_HEALTH | HIDEHUD_PLAYERDEAD | HIDEHUD_NEEDSUIT );
+}
+
+//-----------------------------------------------------------------------------
+// Purpose: 
+//-----------------------------------------------------------------------------
+void CSDKHudArmor::Init( void )
+{
+	Reset();
+	m_iArmor		= INIT_ARMOR;
+}
+
+//-----------------------------------------------------------------------------
+// Purpose: 
+//-----------------------------------------------------------------------------
+void CSDKHudArmor::Reset( void )
+{
+	SetLabelText(L"Armor");
+}
+
+//-----------------------------------------------------------------------------
+// Purpose: 
+//-----------------------------------------------------------------------------
+void CSDKHudArmor::VidInit( void )
+{
+	Reset();
+}
+
+//-----------------------------------------------------------------------------
+// Purpose: 
+//-----------------------------------------------------------------------------
+void CSDKHudArmor::OnThink( void )
+{
+	int newArmor = 0;
+	C_SDKPlayer *local = C_SDKPlayer::GetLocalSDKPlayer();
+	if ( local )
+	{
+		// Never below zero
+		newArmor = MAX( local->GetArmorValue(), 0 );
+	}
+	if ( m_iArmor == newArmor )
+		return;
+
+	if ( !newArmor )
+	{
+	 	GetClientMode()->GetViewportAnimationController()->StartAnimationSequence("SuitPowerZero");
+	}
+	else if ( newArmor < m_iArmor )
+	{
+		// battery power has decreased, so play the damaged animation
+		GetClientMode()->GetViewportAnimationController()->StartAnimationSequence("SuitDamageTaken");
+
+		// play an extra animation if we're super low
+		if ( newArmor < 20 )
+		{
+			GetClientMode()->GetViewportAnimationController()->StartAnimationSequence("SuitArmorLow");
+		}
+	}
+	else
+	{
+		// battery power has increased (if we had no previous armor, or if we just loaded the game, don't use alert state)
+		if ( m_iArmor == INIT_ARMOR || m_iArmor == 0 || newArmor >= 20)
+		{
+			GetClientMode()->GetViewportAnimationController()->StartAnimationSequence("SuitPowerIncreasedAbove20");
+		}
+		else
+		{
+			GetClientMode()->GetViewportAnimationController()->StartAnimationSequence("SuitPowerIncreasedBelow20");
+		}
+	}
+
+	m_iArmor = newArmor;
+
+	//Tony; moved down to fix localization crash at load, update the label name and armor value when it changes.
+	SetLabelText(g_pVGuiLocalize->Find("#SDK_HudArmor"));
+	SetDisplayValue(m_iArmor);
+}
+
+
