--- src/game/client/sdk/sdk_backgroundpanel.h
@@ -1,25 +1,0 @@
-//======== Copyright © 1996-2008, Valve Corporation, All rights reserved. =========//
-//
-// Purpose: 
-//
-// $NoKeywords: $
-//=================================================================================//
-
-#ifndef SDK_BACKHROUNDPANEL_H
-#define SDK_BACKHROUNDPANEL_H
-
-#include <vgui_controls/Frame.h>
-#include <vgui_controls/EditablePanel.h>
-
-//-----------------------------------------------------------------------------
-// Purpose: 
-//-----------------------------------------------------------------------------
-void DrawRoundedBackground( Color bgColor, int wide, int tall );
-
-//-----------------------------------------------------------------------------
-// Purpose: 
-//-----------------------------------------------------------------------------
-void DrawRoundedBorder( Color borderColor, int wide, int tall );
-
-//-----------------------------------------------------------------------------
-#endif // SDK_BACKHROUNDPANEL_H
