--- src/game/client/sdk/sdk_hud_targetid.cpp
@@ -1,222 +1,0 @@
-//========= Copyright © 1996-2005, Valve Corporation, All rights reserved. ============//
-//
-// Purpose: HUD Target ID element
-//
-// $NoKeywords: $
-//=============================================================================//
-#include "cbase.h"
-#include "hud.h"
-#include "hudelement.h"
-#include "c_sdk_player.h"
-#include "c_playerresource.h"
-#include "vgui_EntityPanel.h"
-#include "iclientmode.h"
-#include "vgui/ILocalize.h"
-#include "sdk_gamerules.h"
-
-// memdbgon must be the last include file in a .cpp file!!!
-#include "tier0/memdbgon.h"
-
-#define PLAYER_HINT_DISTANCE	150
-#define PLAYER_HINT_DISTANCE_SQ	(PLAYER_HINT_DISTANCE*PLAYER_HINT_DISTANCE)
-
-static ConVar hud_centerid( "hud_centerid", "1" );
-static ConVar hud_showtargetid( "hud_showtargetid", "1" );
-
-//-----------------------------------------------------------------------------
-// Purpose: 
-//-----------------------------------------------------------------------------
-class CSDKTargetId : public CHudElement, public vgui::Panel
-{
-	DECLARE_CLASS_SIMPLE( CSDKTargetId, vgui::Panel );
-
-public:
-	CSDKTargetId( const char *pElementName );
-	void Init( void );
-	virtual void	ApplySchemeSettings( vgui::IScheme *scheme );
-	virtual void	Paint( void );
-	void VidInit( void );
-
-private:
-	Color			GetColorForTargetTeam( int iTeamNumber );
-
-	vgui::HFont		m_hFont;
-	int				m_iLastEntIndex;
-	float			m_flLastChangeTime;
-};
-
-DECLARE_HUDELEMENT( CSDKTargetId );
-
-using namespace vgui;
-
-//-----------------------------------------------------------------------------
-// Purpose: 
-//-----------------------------------------------------------------------------
-CSDKTargetId::CSDKTargetId( const char *pElementName ) :
-	CHudElement( pElementName ), BaseClass( NULL, "TargetID" )
-{
-	vgui::Panel *pParent = GetClientMode()->GetViewport();
-	SetParent( pParent );
-
-	m_hFont = g_hFontTrebuchet24;
-	m_flLastChangeTime = 0;
-	m_iLastEntIndex = 0;
-
-	SetHiddenBits( HIDEHUD_MISCSTATUS );
-}
-
-//-----------------------------------------------------------------------------
-// Purpose: Setup
-//-----------------------------------------------------------------------------
-void CSDKTargetId::Init( void )
-{
-};
-
-void CSDKTargetId::ApplySchemeSettings( vgui::IScheme *scheme )
-{
-	BaseClass::ApplySchemeSettings( scheme );
-
-	m_hFont = scheme->GetFont( "TargetID", IsProportional() );
-
-	SetPaintBackgroundEnabled( false );
-}
-
-//-----------------------------------------------------------------------------
-// Purpose: clear out string etc between levels
-//-----------------------------------------------------------------------------
-void CSDKTargetId::VidInit()
-{
-	CHudElement::VidInit();
-
-	m_flLastChangeTime = 0;
-	m_iLastEntIndex = 0;
-}
-
-Color CSDKTargetId::GetColorForTargetTeam( int iTeamNumber )
-{
-	return GameResources()->GetTeamColor( iTeamNumber );
-} 
-
-//-----------------------------------------------------------------------------
-// Purpose: Draw function for the element
-//-----------------------------------------------------------------------------
-void CSDKTargetId::Paint()
-{
-#define MAX_ID_STRING 256
-	wchar_t sIDString[ MAX_ID_STRING ];
-	sIDString[0] = 0;
-
-	C_SDKPlayer *pPlayer = C_SDKPlayer::GetLocalSDKPlayer();
-
-	if ( !pPlayer )
-		return;
-
-	Color c;
-
-	// Get our target's ent index
-	int iEntIndex = pPlayer->GetIDTarget();
-	// Didn't find one?
-	if ( !iEntIndex )
-	{
-		// Check to see if we should clear our ID
-		if ( m_flLastChangeTime && (gpGlobals->curtime > (m_flLastChangeTime + 0.5)) )
-		{
-			m_flLastChangeTime = 0;
-			sIDString[0] = 0;
-			m_iLastEntIndex = 0;
-		}
-		else
-		{
-			// Keep re-using the old one
-			iEntIndex = m_iLastEntIndex;
-		}
-	}
-	else
-	{
-		m_flLastChangeTime = gpGlobals->curtime;
-	}
-
-	// Is this an entindex sent by the server?
-	if ( iEntIndex )
-	{
-		C_BasePlayer *pPlayer = static_cast<C_BasePlayer*>(cl_entitylist->GetEnt( iEntIndex ));
-		C_BasePlayer *pLocalPlayer = C_BasePlayer::GetLocalPlayer();
-
-		const char *printFormatString = NULL;
-		wchar_t wszPlayerName[ MAX_PLAYER_NAME_LENGTH ];
-		wchar_t wszHealthText[ 10 ];
-		bool bShowHealth = false;
-		bool bShowPlayerName = false;
-
-		// Some entities we always want to check, cause the text may change
-		// even while we're looking at it
-		// Is it a player?
-		if ( IsPlayerIndex( iEntIndex ) )
-		{
-			c = GetColorForTargetTeam( pPlayer->GetTeamNumber() );
-
-			bShowPlayerName = true;
-			g_pVGuiLocalize->ConvertANSIToUnicode( pPlayer->GetPlayerName(),  wszPlayerName, sizeof(wszPlayerName) );
-			
-			if ( SDKGameRules()->IsTeamplay() == true && pPlayer->InSameTeam(pLocalPlayer) )
-			{
-				printFormatString = "#SDK_Playerid_sameteam";
-				bShowHealth = true;
-			}
-			else
-			{
-				printFormatString = "#SDK_Playerid_diffteam";
-			}
-		
-
-			if ( bShowHealth )
-			{
-				_snwprintf( wszHealthText, ARRAYSIZE(wszHealthText) - 1, L"%.0f%%",  ((float)pPlayer->GetHealth() / (float)pPlayer->GetMaxHealth() ) );
-				wszHealthText[ ARRAYSIZE(wszHealthText)-1 ] = '\0';
-			}
-		}
-
-		if ( printFormatString )
-		{
-			if ( bShowPlayerName && bShowHealth )
-			{
-				g_pVGuiLocalize->ConstructString( sIDString, sizeof(sIDString), g_pVGuiLocalize->Find(printFormatString), 2, wszPlayerName, wszHealthText );
-			}
-			else if ( bShowPlayerName )
-			{
-				g_pVGuiLocalize->ConstructString( sIDString, sizeof(sIDString), g_pVGuiLocalize->Find(printFormatString), 1, wszPlayerName );
-			}
-			else if ( bShowHealth )
-			{
-				g_pVGuiLocalize->ConstructString( sIDString, sizeof(sIDString), g_pVGuiLocalize->Find(printFormatString), 1, wszHealthText );
-			}
-			else
-			{
-				g_pVGuiLocalize->ConstructString( sIDString, sizeof(sIDString), g_pVGuiLocalize->Find(printFormatString), 0 );
-			}
-		}
-
-		if ( sIDString[0] )
-		{
-			int wide, tall;
-			int ypos = YRES(260);
-			int xpos = XRES(10);
-
-			vgui::surface()->GetTextSize( m_hFont, sIDString, wide, tall );
-
-			if( hud_centerid.GetInt() == 0 )
-			{
-				ypos = YRES(420);
-			}
-			else
-			{
-				xpos = (ScreenWidth() - wide) / 2;
-			}
-			
-			vgui::surface()->DrawSetTextFont( m_hFont );
-			vgui::surface()->DrawSetTextPos( xpos, ypos );
-			vgui::surface()->DrawSetTextColor( c );
-			vgui::surface()->DrawPrintText( sIDString, wcslen(sIDString) );
-		}
-	}
-}
